walt-server-setup: make it restartable if apt upgrade fails
* Record in a file /var/lib/walt/.upgrade the OS version
  which we are upgrading from.
* At the end of the upgrade, remove this file.
* Ensure the venv is always upgraded to the possibly new
  python version, even if the apt package upgrade failed.

This ensures that in case of apt package upgrade failure,
after correcting the issue (with the help of the upgrade
log file) the user can retry walt-server-setup. The file
/var/lib/walt/.upgrade being present means the previous
run failed to complete, so a full-upgrade process is
restarted.

Note: if the python package was properly updated but another
package caused the issue later, it is mandatory that the
venv is upgraded (3rd step) since it embeds `walt-server-setup`
which was linked to the previous python interpreter.