Fix changing netsetup on virtual nodes.
Changing `netsetup` configuration of a virtual node for allowing or
disabling internet access was not applied.
The user had to reboot the server or restart the main service as root,
i.e., `systemctl restart walt-server` for this change to be applied.

Actually the NAT rules were properly updated but the virtual node
manager was not notified of the new "gateway" setting.
This commit adds this notification.

Fix #118.