server/logs: apply logline and stream name regex filters in sql query
When retrieving logs from db (i.e. using `--history`), we previously
applied the optional regex on stream name and/or log line in the
python code, because the format of regular expressions was different
in python and in postgresql, and we had to maintain coherency with
realtime filtering (i.e. using `--realtime`).

Since we now have implemented posix-compliant extended regular
expressions (cf. previous commit), and we can limit the syntax
in postgresql to the same format by using regex prefix "(?e)",
we now apply them on the database queries.

This improves the performance for requests implying those
regular expressions.