server: avoid using blocking process to stream db logs
The blocking process is sometimes busy with blocking operations
such as trying to send SNMP requests to a switch which is no longer
accessible.

The procedure to retrieve logs from db was previously relying on
the blocking process, sometimes leading to hangups of several seconds
if such situations.

We now fully handle this procedure in the main process, by using
a Workflow object and async calls to the db process.