nfs service: Tune ExecReload to avoid a whole service restart
"systemctl restart nfs-kernel-server" is quite long, but we were
still using this command. "systemctl reload" is faster, but it
did not remove exports currently is use by some clients, which
could prevent the related images to be unmounted.

This commit
* adds an override to the systemd unit file to add a
call to "exportfs -f", which allows to flush those exports
* uses "systemctl reload" instead of "systemctl restart",
resulting in a much faster operation.