Workaround issue with podman unix socket and http proxies
walt-server-daemon uses a unix socket and a PodmanClient
object to communicate with walt-server-podman service.

This object internally uses library "requests", and if we
define an HTTP proxy in environment variables, this library
wrongly redirects the connection, even if we are targetting
a UNIX socket.

This commit works around this issue by using a setting
telling "requests" not to trust the environment.