nbfsd: use the new unix socket to request reloading conf.
Previously, we were just sending a SIGHUP signal.
The new method of sending "RELOAD_CONF" on the unix socket
at /run/nbfsd/ctl.sock is better because we can wait for
the response (should be "OK") before continuing, so we are
sure that nbfsd is no longer holding any open files on the
exports no longer accepted.