walt-image-umount: loop if some processes prevent immediate umount
When walt-server-daemon is interrupted, umounting images was sometimes
failing because some processes such as in.tftpd, nbfsd or maybe nfsd where
still serving the content of a file to a node.
We now retry in a loop, and print the PIDs of processes involved if we
can detect them by command `lsof`.