server: fix rare bug related to get_username() caching
The bug was visible when .walt/config was missing hub
credentials and we were running "walt advanced update-hub-meta".
In this case, credentials were prompted but then the server
was crashing because of the main process was sending the
get_username() request to the hub process with an obsolete
req_id that the hub process did not find.

It was related to the fact caching was occuring in the
APISession object, and this object may be shared for
several consecutive tasks. When there was subtle
reentrance scenarios, this could lead to use the RPCContext
of a task already finished instead of the current one.