server: avoid reentrance on blocking process for registry creds
In the current code, if the blocking process was missing registry
credentials, it was calling back the requester on the client.
If thos credentials were missing in the client conf, they were
prompted to the user, and then checked by trying a login() operation.
This login() was calling the blocking process again, so this was
a reentrance case.

After we enforced the main->blocking requests to be done one at
a time, this was causing a deadlock.

This commit replies with a failing code when this occurs. This
code is handled by main, which request the hub to prompt missing
credentials from the user, and then retries the operation.

The main process is reentrant so the concurrent login() operation
for testing credentials no longer is a problem.