wip -- not working -- serve-local-dns
The purpose here is to query walt-server-named for DNS requests
issued locally on the server, e.g., `ping <walt-device>`.

This code tries to update /etc/resolv.conf appropriately when
walt-server-daemon starts; but the dhcp response on the default
interface arrives a few seconds later on bootup, so the file
is overwritten.

We give up on this for now, and it could possibly be implemented
successfully later, if we use systemd-resolved and systemd-networkd,
possibly configuring the specific DNS information on walt-net
interface only, or by better tuning the order of the services (see
https://stackoverflow.com/a/54300998/4733396).