hybrid-p mode: remove the content of /tmp
walt-net-service creates fifos in /tmp and is supposed to run as
an infinite loop, so these files are never removed.
Their name contain a PID number, but when rebooting the PID number
may be the same as a previous run, and mkfifo failed because the
fifo already existed.
Instead of removing the fifo if it already exists, this commit
removes the whole tmp directory of the diff directory before mounting
the overlay. This is a more generic fix, just in case other cases
like this could occur in other scripts.