server: let nodes report their lldp position
When using up-to-date server software and up-to-date
debian images, nodes run a systemd service based on
"lldpcli watch" in order to report LLDP neighbors to
the server.

The server exposes a script at
[image-root]:/bin/walt-report-lldp-neighbor
for this purpose.
This script takes the neighbor (i.e. switch) mac
and port as parameters. The port is given as an
LLDP label.
The server maintains a cache for the association
between the LLDP label and the port number.
If this information is not known yet, then a rescan
of the switch is triggered. It the switch supports
LLDP information retrieval, then all LLDP port labels
of the switch are saved in the cache. Once those
port labels are known, no rescans are needed next time
this switch if concerned by a neighbor report.
If the switch does not support LLDP information retrieval,
then the mac of the node is looked up in the topology
results (deduced from the BRIDGE table), and the
corresponding port number is associated with the LLDP port
label of the initial request.
In this case, neighbor reports of nodes connected on other
ports of this switch will need one more rescan, but only
the first time for each port.

Fix #125.