walt-init of vpn nodes: fix missing mv symlink in rootfs
The bug was previously not visible because the mv symlink
was finally still created when detecting the boot-mode.
However, due to the recent commit(*), this boot-mode
detection code is no longer called for VPN nodes (because this
commit disallows hybrid boot modes for them for security reason,
so the only boot mode possible for them is network boot).

This commit properly adds 'mv' in the list of needed
busybox applets in walt-init.

(*) commit "Security: disallow hybrid boot for VPN nodes."