walt with no argument: print server and auto-completion status
When typing `walt` with no arguments, a "status box" is now printed
to display the status of the client installation:
* the version (this was previously the only thing displayed there)
* the server IP or hostname
* the completion status (i.e., up-to-date or not).

The completion status can be detected because we now export one new
variable called _WALT_COMP_VERSION in the shell completion scripts
generated by `walt advanced dump-bash-autocomplete` and
`walt advanced dump-zsh-autocomplete`.
See #135 for the original idea.

Two other variables are also added, for possible future use:
* `_WALT_COMP_TYPE`, with value `bash` or `zsh`
* `_WALT_COMP_BASH_SOURCE` or `_WALT_COMP_ZSH_SOURCE`, giving the
  source path of the completion script.

The logo printing code was updated because we now use a few escape
codes for displaying this status box, so the length of some of the
text lines is no longer the number of bytes.

The documentation about shell completion was updated.