Detect and mount local swap on nodes
This is done in walt-init.

If both local swap (disk based) and remote swap (NBD client) are
available, only local swap is set up because it is more reliable.

However, local swap is disabled for VPN-capable nodes
(Raspberry Pi 5 boards).

A new help topic called 'node-swap' was created and a part of
the 'boot-modes' topic was offloaded there.